app/requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.13
pytz==2023.3
Requests==2.31.0
